+++ collector/src/main/java/com/bitmovin/analytics/data/DeviceClass.kt
@@ -1,8 +1,6 @@
 package com.bitmovin.analytics.data
 
-import androidx.annotation.Keep
-
-@Keep // Protect from obfuscation in case customers are using proguard, enums might not be strictly needed but there could be rules that mess up with our dtos
+// DTO which is protected from ProGuard obfuscation through proguard-consumer-rules.pro
 enum class DeviceClass {
     TV,
     Phone,