+++ collector/src/main/java/com/bitmovin/analytics/data/DownloadSpeedInfo.kt
@@ -1,8 +1,6 @@
 package com.bitmovin.analytics.data
 
-import androidx.annotation.Keep
-
-@Keep // Protect from obfuscation in case customers are using proguard
+// DTO which is protected from ProGuard obfuscation through proguard-consumer-rules.pro
 class DownloadSpeedInfo(
     // Number of completed segment downloads
     val segmentsDownloadCount: Int = 0,