+++ collector/src/main/java/com/bitmovin/analytics/data/ErrorCode.kt
@@ -1,11 +1,9 @@
 package com.bitmovin.analytics.data
 
-import androidx.annotation.Keep
 import com.bitmovin.analytics.features.errordetails.ErrorData
 
-@Keep // Protect from obfuscation in case customers are using proguard
+// DTO which is protected from ProGuard obfuscation through proguard-consumer-rules.pro
 data class ErrorCode(val errorCode: Int, val description: String, val errorData: ErrorData, val legacyErrorData: LegacyErrorData? = null) {
-
     @Override
     override fun toString(): String {
         return "$errorCode: $description"