+++ collector/src/main/java/com/bitmovin/analytics/data/LegacyErrorData.kt
@@ -1,7 +1,5 @@
 package com.bitmovin.analytics.data
 
-import androidx.annotation.Keep
-
 // LegacyErrorData is stored unindexed as error_data in crate
-@Keep // Protect from obfuscation in case customers are using proguard
+// DTO which is protected from ProGuard obfuscation through proguard-consumer-rules.pro
 data class LegacyErrorData(val msg: String, val details: List<String> = emptyList())