+++ collector/src/main/java/com/bitmovin/analytics/data/LicenseCallData.kt
@@ -1,6 +1,4 @@
 package com.bitmovin.analytics.data
 
-import androidx.annotation.Keep
-
-@Keep // Protect from obfuscation in case customers are using proguard
+// DTO which is protected from ProGuard obfuscation through proguard-consumer-rules.pro
 data class LicenseCallData(val key: String, val analyticsVersion: String, val domain: String)