+++ collector/src/main/java/com/bitmovin/analytics/data/LicenseResponse.kt
@@ -1,7 +1,6 @@
 package com.bitmovin.analytics.data
 
-import androidx.annotation.Keep
 import com.bitmovin.analytics.license.FeatureConfigContainer
 
-@Keep // Protect from obfuscation in case customers are using proguard
+// DTO which is protected from ProGuard obfuscation through proguard-consumer-rules.pro
 data class LicenseResponse(val status: String?, val message: String?, val features: FeatureConfigContainer?)