+++ collector/src/main/java/com/bitmovin/analytics/features/errordetails/ErrorDetailTrackingConfig.kt
@@ -1,7 +1,6 @@
 package com.bitmovin.analytics.features.errordetails
 
-import androidx.annotation.Keep
 import com.bitmovin.analytics.features.FeatureConfig
 
-@Keep // Protect from obfuscation in case customers are using proguard
+// DTO which is protected from ProGuard obfuscation through proguard-consumer-rules.pro
 data class ErrorDetailTrackingConfig(override val enabled: Boolean = false, val numberOfHttpRequests: Int? = null) : FeatureConfig