+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Development
 
+### Changed
+- Use consumer pro guard rules defined in proguard-consumer-rules.pro instead of @Keep to prevent obfuscation of DTOs
+- Added explicit runtime dependency to kotlin-parcelize-runtime to fix R8 issues, and changed publish script to also add runtime deps to pom.xml
+
 ## v3.12.0
 
 ### Fixed